Add files via upload
The SEC sued LBRY in March 2021, alleging that LBC tokens were securities and that the startup had violated securities laws by selling them without registering with the agency. LBRY pushed back, claiming that LBC tokens were not securities, and that the SEC did not give it fair notice that its sale of LBC was subject to securities laws, thus violating the company’s right to due process. The LBRY protocol is a blockchain-based file-sharing network.
One Nov 7, 2022, Federal Judge Paul Barbadoro, of the District Court for the District of New Hampshire, ruled that “no reasonable trier of fact could reject the SEC’s contention that LBRY offered LBC as a security, and LBRY does not have a triable defense that it lacked fair notice.”
LBRY founder Jeremy Kauffman has long maintained that the outcome of the case could have sweeping implications for the wider crypto industry. Speaking at Messari’s Mainnet conference in September, Kauffman said the facts in LBRY’s case would “basically apply to every company” in the crypto industry.
"The SEC vs LBRY case establishes a precedent that threatens the entire US cryptocurrency industry. Under the SEC vs LBRY standard, almost every cryptocurrency, including Ethereum and Doge, are securities. The future of cryptocurrency in the US now rests in with an organization even worse than the SEC: the United States Congress," he said.